Session的Error改成int类型，方便自己增加错误码，并且在Session中增加Error字段，方便 打印错误码，ActorMessageSender也增加了Error字段

+++ Unity/Assets/Scripts/Module/Message/Network/KCP/KChannel.cs
@@ -153,7 +153,7 @@ namespace ETModel
 			// 超时断开连接
 			if (timeNow - this.lastRecvTime > 20 * 1000)
 			{
-				this.OnError(SocketError.Disconnecting);
+				this.OnError((int)SocketError.Disconnecting);
 				return;
 			}
 			this.kcp.Update(timeNow);
@@ -185,7 +185,7 @@ namespace ETModel
 				int n = kcp.PeekSize();
 				if (n == 0)
 				{
-					this.OnError(SocketError.NetworkReset);
+					this.OnError((int)SocketError.NetworkReset);
 					return;
 				}
 				int count = this.kcp.Recv(this.cacheBytes);
